backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dateutil==2.8.2
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
scikit-learn==1.3.2
joblib==1.3.2
